refactor : add domain and view package
