+++ src/main/java/racing/domain/RaceResult.java
@@ -1,4 +1,4 @@
-package racing.race;
+package racing.domain;
 
 import java.util.Comparator;
 import java.util.List;
