+++ src/main/java/racing/dto/RaceAttempt.java
@@ -1,4 +1,4 @@
-package racing.input;
+package racing.dto;
 
 import racing.exception.InvalidInputValueException;
