+++ src/main/java/racing/view/input/InputView.java
@@ -1,4 +1,6 @@
-package racing.input;
+package racing.view.input;
+
+import racing.dto.RaceInfo;
 
 import java.util.InputMismatchException;
 import java.util.Scanner;
@@ -7,13 +9,13 @@ public class InputView {
 
     private static final Scanner SCANNER = new Scanner(System.in);
 
-    public RaceInput racingInput() {
+    public RaceInfo racingInput() {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         String carInfo = getString();
         System.out.println("시도할 횟수는 몇 회 인가요?");
         int numOfAttempt = getInt();
 
-        return new RaceInput(carInfo, numOfAttempt);
+        return new RaceInfo(carInfo, numOfAttempt);
     }
 
     private String getString() {
