+++ src/test/java/racing/domain/RandomNumRaceRuleTest.java
@@ -1,6 +1,5 @@
-package racing.race;
+package racing.domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
