+++ src/test/java/racing/dto/RaceLineupTest.java
@@ -1,4 +1,4 @@
-package racing.input;
+package racing.dto;
 
 import org.junit.jupiter.api.Test;
